Move exection engine to separate file
